requirements.txt
#data processing
pandas
openpyxl

#env mgmt
python-dotenv

#google apis
google-api-core==2.3.2
google-api-python-client==2.33.0
google-cloud-bigquery==3.1.0
google-cloud-bigquery[pandas] # need for db-dtypes==1.0.1

